backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.2
tqdm==4.64.1
torchsummary==1.5.1
transformers==4.21.1
pyyaml==6.0
lorem==0.1.1
torchinfo==1.8.0
gradio==4.12.0